+++ main_env_var_local_test.go
@@ -10,7 +10,7 @@ import (
 
 func TestEnvLocalVarCreate(t *testing.T) {
 	t.Parallel()
-	updateGolden := os.Getenv("envelope_TEST_UPDATE_GOLDEN") != ""
+	updateGolden := os.Getenv("ENVELOPE_TEST_UPDATE_GOLDEN") != ""
 
 	dbFile, err := os.CreateTemp(os.TempDir(), "envelope-test-")
 	require.NoError(t, err)
@@ -91,7 +91,7 @@ func TestEnvLocalVarCreate(t *testing.T) {
 
 func TestEnvLocalVarDelete(t *testing.T) {
 	t.Parallel()
-	updateGolden := os.Getenv("envelope_TEST_UPDATE_GOLDEN") != ""
+	updateGolden := os.Getenv("ENVELOPE_TEST_UPDATE_GOLDEN") != ""
 
 	dbFile, err := os.CreateTemp(os.TempDir(), "envelope-test-")
 	require.NoError(t, err)